+++ keyboards/nuphy/gem80/ansi/debounce.c
@@ -7,7 +7,7 @@ releasing a key, that state is pushed after no changes occur for DEBOUNCE millis
 #include "debounce.h"
 #include "timer.h"
 #include <stdlib.h>
-#include "ansi.h"
+#include "gem80-common.h"
 
 #ifdef PROTOCOL_CHIBIOS
 #    if CH_CFG_USE_MEMCORE == FALSE
