+++ keyboards/nuphy/gem80/ansi/gem80-common.h
