+++ keyboards/nuphy/gem80/ansi/rules.mk
@@ -2,7 +2,7 @@ SRC += user_kb.c
 SRC += rf.c
 SRC += side.c side_driver.c side_logo.c
 
-# can be separated with needed fields implemented in ansi.h
+# can be separated with needed fields implemented in gem80-common.h
 SRC += mcu_pwr.c sleep.c debounce.c rf_driver.c rf_queue.c
 
 SRC += gem80-common.c
