+++ keyboards/nuphy/gem80/ansi/side_logo.c
@@ -1,6 +1,6 @@
 // Copyright 2023 Persama (@Persama)
 // SPDX-License-Identifier: GPL-2.0-or-later
-#include "ansi.h"
+#include "gem80-common.h"
 #include "rgb_matrix.h"
 #include "side.h"
 #include "timer.h"
